Ajout d'une table relationnel entre service et site

+++ Server/AnnuaireServices/ServiceServices.cs/ServiceService.cs
@@ -1,17 +1,22 @@
 using Microsoft.EntityFrameworkCore;
 using BlazorAnnuaireProject.Context;
 using BlazorAnnuaireProject.Entities;
+using BlazorAnnuaireProject.Models;
+using AutoMapper;
+using BlazorAnnuaireProject.Helpers;
 
 namespace BlazorAnnuaireProject.AnnuaireServices.ServiceService;
 
 public class ServiceService : IServiceService
 {
     private readonly DataContext _context;
-
-    public ServiceService(DataContext context)
+    private readonly IMapper _mapper;
+    public ServiceService(DataContext context, IMapper mapper)
     {
         _context = context;
+        _mapper = mapper;
     }
+
 
     public async Task<List<Services>> GetAllServices()
     {
@@ -23,24 +28,89 @@ public class ServiceService : IServiceService
         return await _context.Services.Include(s => s.Salaries).FirstOrDefaultAsync(s => s.Id == id);
     }
 
-    public async Task<int> CreateService(Services service)
+    public CreateServiceResponse CreateService(CreateServiceRequest service)
     {
-        _context.Services.Add(service);
-        await _context.SaveChangesAsync();
-        return service.Id;
+        var registerService = _mapper.Map<Services>(service);
+        _context.Services.Add(registerService);
+        _context.SaveChanges();
+        return new CreateServiceResponse(registerService);
     }
 
-    public async Task UpdateService(Services service)
+    // public CreateServiceResponse CreateServiceOnSite(CreateServiceRequest service, string Ville)
+    // {
+    //     var site = _context.Sites.FirstOrDefault(s => s.Ville == Ville);
+    //     if (site == null)
+    //     {
+    //         throw new Exception("Site not found");
+    //     }
+    //     // if (site.Services.Any(s => s.Nom == service.Nom))
+    //     // {
+    //     //     throw new Exception("Service already exists");
+    //     // }
+
+    //     // var registerService = _mapper.Map<Services>(service);
+    //     // site.Services.Add(registerService);
+    //     _context.Entry(site).State = EntityState.Modified;
+    //     _context.SaveChanges();
+    //     return new CreateServiceResponse();
+    // }
+
+    public UpdateServiceResponse UpdateService(UpdateServiceRequest service, string nom)
     {
-        _context.Entry(service).State = EntityState.Modified;
-        await _context.SaveChangesAsync();
+        var existingService = _context.Services.FirstOrDefault(s => s.Nom == nom);
+        if (existingService == null)
+        {
+            throw new Exception("Service not found");
+        }
+        if (SalariesOnServices(existingService.Id))
+        {
+            throw new Exception("Service has salaries");
+        }
+
+        bool updated = false;
+        if (!string.IsNullOrWhiteSpace(service.Nom) && !string.IsNullOrEmpty(service.Nom) && service.Nom != existingService.Nom && service.Nom != "string")
+        {
+            existingService.Nom = service.Nom;
+            updated = true;
+        }
+
+        if (updated)
+        {
+            _context.Entry(existingService).State = EntityState.Modified;
+            _context.SaveChanges();
+            return new UpdateServiceResponse(existingService, "Service updated");
+        }
+
+        return new UpdateServiceResponse(existingService, "Service not updated");
+
     }
 
-    public async Task DeleteService(int id)
+    public DeleteServiceResponse DeleteService(string nom)
     {
-        var service = await _context.Services.FindAsync(id);
+        var service = _context.Services.FirstOrDefault(s => s.Nom == nom);
+        if (service == null)
+        {
+            throw new Exception("Service not found");
+        }
+        if (SalariesOnServices(service.Id))
+        {
+            throw new Exception("Service has salaries");
+        }
+        var infoRemoved = service.Nom;
         _context.Services.Remove(service);
-        await _context.SaveChangesAsync();
+        _context.SaveChangesAsync();
+        return new DeleteServiceResponse(infoRemoved, "Service deleted");
     }
+
+    private bool SalariesOnServices(int serviceId)
+    {
+        var site = _context.Services.Include(s => s.Salaries).FirstOrDefault(s => s.Id == serviceId);
+        if (site != null && site.Salaries.Count > 0)
+        {
+            return true;
+        }
+        return false;
+    }
+
 }
